refactor(editor): drop redundant theme dispatch + isolate wiring tests (#114)
